# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "EPIC mail",
    "description": "This is an API documentation for a web aplication named EPIC mail. It will cater for the Messaging needs of users who sign up on it"
  },
  "schemes": [
    "https"
  ],
  "host": "epic-mail-application.herokuapp.com",
  "basePath": "/api/v1",
  "paths": {
...
    "/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Gets Authorization",
        "description": "Signs in User.",
        "parameters": [
          {
            "name": "user",
            "in": "body",
            "description": "The user to Authorize.",
            "schema": {
              "$ref": "#/definitions/SendAuthData"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User succesfully sign in and receives a token.",
            "schema": {
              "$ref": "#/definitions/returnNewUserData"
            }
          },
          "400": {
            "description": "Invalid request.",
            "schema": {
              "$ref": "#/definitions/returnError"
            }
          }
        }
      }
    },
    "/auth/signup": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Gets Authorization",
        "description": "Signs in User.",
        "parameters": [
          {
            "name": "user",
            "in": "body",
            "description": "The user to Authorize.",
            "schema": {
              "$ref": "#/definitions/SendNewUserData"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User succesfully sign in and receives a token.",
            "schema": {
              "$ref": "#/definitions/returnNewUserData"
            }
          },
          "400": {
            "description": "Invalid request.",
            "schema": {
              "$ref": "#/definitions/returnError"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "SendNewUserData": {
      "type": "object",
      "properties": {
        "firstName": {
          "type": "string"
        },
        "lastName": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "recoveryEmail": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "confirmPassword": {
          "type": "string"
        },
        "phoneNumber": {
          "type": "string"
        }
      },
      "xml": {
        "name": "Tag"
      }
    },
    "SendAuthData": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      },
      "xml": {
        "name": "Tag"
      }
    },
...
    "returnError": {
      "type": "object",
      "required": [
        "status",
        "error"
      ],
      "properties": {
        "status": {
          "type": "integer",
          "format": "int64"
        },
        "error": {
          "type": "string"
        }
      }
    },
    "NewUserData": {
      "type": "object",
      "properties": {
        "token": {
          "type": "string"
        },
        "firstname": {
          "type": "string"
        },
        "lastname": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "phonenumber": {
          "type": "string"
        },
        "recoveryemail": {
          "type": "string"
        },
        "id": {
          "type": "number"
        },
        "dp": {
          "type": "string"
        }
      },
      "xml": {
        "name": "NewUserData"
      }
    },
    "returnNewUserData": {
      "type": "object",
      "required": [
        "status",
        "data"
      ],
      "properties": {
        "status": {
          "type": "integer",
          "format": "int64"
        },
        "data": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/NewUserData"
          }
        }
      }
    }
  }
}